Move const string with non ascii char to language file.

+++ Common/Source/Library/poco/TextConverter.cpp
@@ -77,7 +77,7 @@ int TextConverter::convert(const std::string& source, std::string& destination, 
 		c = trans(c);
 		int n = _outEncoding.convert(c, buffer, sizeof(buffer));
 		if (n == 0) n = _outEncoding.convert(_defaultChar, buffer, sizeof(buffer));
-		poco_assert (n <= sizeof(buffer));
+		poco_assert ((size_t)n <= sizeof(buffer));
 		destination.append((const char*) buffer, n);
 		++it;
 	}
@@ -128,7 +128,7 @@ int TextConverter::convert(const void* source, int length, std::string& destinat
 		uc = trans(uc);
 		n = _outEncoding.convert(uc, buffer, sizeof(buffer));
 		if (n == 0) n = _outEncoding.convert(_defaultChar, buffer, sizeof(buffer));
-		poco_assert (n <= sizeof(buffer));
+		poco_assert ((size_t)n <= sizeof(buffer));
 		destination.append((const char*) buffer, n);
 	}
 	return errors;
